fix the way gravity influence acceleration
previously i accumulated the gravitation force to UPDATE
acceleration. in this version i fixed the issue by using the
gravitational force to SET the acceleration.

i also think the old version isn't exactly wrong - i used the term
"acceleration" in the code but in fact it was used for the change in
position and accumulation of gravitation force, so it's a
wrong use of term instead of an actual physical error.